♻️Refactor : color 테이블에서 수행되는 메소드 정의 수정
