+++ src/main/java/pickRAP/server/repository/color/ColorRepositoryCustom.java
@@ -13,4 +13,5 @@ public interface ColorRepositoryCustom {
 
     List<PersonalMoodResponse> getPersonalMoodAnalysisResults(Member member);
 
+    List<Magazine> findTop20MagazineByColor();
 }
