+++ src/main/java/pickRAP/server/repository/magazine/MagazineRepositoryCustom.java
@@ -8,6 +8,5 @@ public interface MagazineRepositoryCustom {
     List<Magazine> findMemberMagazines(String email);
     List<Magazine> findMagazineByHashtag(String keyword);
     List<Magazine> findMagazineByHashtagAndNotWriter(List<String> keyword, String email);
-    List<Magazine> findTop20MagazineByColor();
 
 }
